Separation of Concerns
Created functioning test.py that basically adds separation of concerns. Controller shouldn't be worried about what the view and models do so the test.py file is an example to work off of. If we want to modify a label for example, we can do what's in this file and add more functions to keep modifying stuff